services/web/requirements.txt
Flask==3.0.0
flask-limiter==3.5.0
flask-login==0.6.3
flask-mailman==1.0.0
flask-security-too==5.3.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
SQLAlchemy==2.0.23
WTForms==3.1.1
zxcvbn==4.4.28
bleach==6.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
cryptography==41.0.7
qrcode==7.4.2